==> Scripts/5. Deriving New Variables.ipynb <==
### This script contains the following points:

#### 1. Importing libraries
#### 2. Importing orders_products_merged
#### 3. Create price_range column and labels for product prices
#### 4. Create price_label and busiest_day columns
#### 5. Create new column for busiest_period_of_day
#### 6. Exporting ords_prods_merge to pkl

# 1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 2. Importing orders_products_merged

# importing PKL order_products_combined from prepared data
ords_prods_merge = pd.read_pickle(r'C:\Users\Documents\Career Foundry\Data Immersion\Instacart Basket Analysis\02 Data\Prepared Data\orders_products_merged.pkl')

# assigning main project path to variable 'path'
path = r'C:\Users\Documents\Career Foundry\Data Immersion\Instacart Basket Analysis'

# 3. Create price_range column and labels for product prices

#subset of dataset to conserve memory
df = ords_prods_merge[:1000000]

df.shape

# creating a function to assign labels to products based on price

def price_label(row):

    if row['prices'] <= 5:
        return 'Low-range product'
    elif (row['prices'] > 5) and (row['prices'] <= 15):
        return 'Mid-range product'
    elif row['prices'] > 15:
        return 'High range'
    else: return 'Not enough data'

#apply the function by filling values in new price_range column

df['price_range'] = df.apply(price_label, axis=1)

#counts of values in different ranges for first 1 million rows

df['price_range'].value_counts(dropna = False)

#finding the max price in subset, result confirms that there are no high range items in the subset

df['prices'].max()

# 4. Create price_label and busiest_day columns to ords_prods_merge

# checking shape for reference
ords_prods_merge.shape

# similar to function in section 3, but faster
#filters are applied before searching data frame (rather than search then apply for if/then)

ords_prods_merge.loc[ords_prods_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'

ords_prods_merge.loc[(ords_prods_merge['prices'] <= 15) & (ords_prods_merge['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

ords_prods_merge.loc[ords_prods_merge['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

#able to search entire df after using loc() function

ords_prods_merge['price_range_loc'].value_counts(dropna = False)

### Counts of orders for each day of week. Saturday (0) is busiest and Wednesday (4) is slowest.

# number of orders by day of week
ords_prods_merge['orders_day_of_week'].value_counts(dropna = False)

### Using If-statements to create labels for the busiest days (0 & 1), slowest days (3 & 4) and regularly busy days (2,5,6)

# condition will create busiest_day column and apply labels to each day of week

result = []

for value in ords_prods_merge["orders_day_of_week"]:
    if value <= 1:
        result.append("Busiest")
    elif value == 4 and value == 3:
        result.append("Slowest")
    else:
        result.append("Regularly busy")
ords_prods_merge["busiest_day"] = result
ords_prods_merge

#counts of each category which match the counts of DOW calculated before
ords_prods_merge['busiest_day'].value_counts(dropna = False)

# checking shape for reference
ords_prods_merge.shape

#successful addition of new column and new results
ords_prods_merge.head()

# 5.  Create new column for busiest_period_of_day

# busiest hours of day: most orders, average, fewest
# 24 hours of day divided by three time frame labels
#counts for orders per hour of day calculated and sorted by most orders

ords_prods_merge.value_counts(['order_hour_of_day'])

# hours greater than or equal to 9 are Most orders
# hours 7, 8, and 17-22 are average orders
# all other hours are fewest orders

result = []

for value in ords_prods_merge["order_hour_of_day"]:
    if value in [9, 10, 11, 12, 13, 14, 15, 16]:
        result.append("Most orders")
    elif value in [17, 8, 18, 19, 20, 7, 21, 22]:
        result.append("Average orders")
    else:
        result.append("Fewest orders")
ords_prods_merge["busiest_period_of_day"] = result
ords_prods_merge

# confirm same number of rows. addition of one new column

ords_prods_merge.shape

# confirm that all values are assigned labels

ords_prods_merge.value_counts(['busiest_period_of_day'])

# 6. Exporting ords_prods_merge to pkl

# Export data to pkl

ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
