==> Scripts/7. Final Merge of Data Sets.ipynb <==
### This script contains the following points:

#### 1. Importing libraries
#### 2. Importing customers data
#### 3. Data Wrangling
#### 4. Renaming customer columns
#### 5. Checking and changing customers data types
#### 6. Checking for nulls and duplicate values
#### 7. Exporting customers df
#### 8. Import ords_prods_merge dataframe
#### 9. Addressing outliers in prices column
#### 10. Change ords_prods_merge data types
#### 11. Merging ords_prods_merge to customers
#### 12. Exporting full_merge to pkl

# 1. Importing libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# 2. Importing customers data

customers = pd.read_csv(r'C:\Users\Documents\Career Foundry\Data Immersion\Instacart Basket Analysis\02 Data\Original Data\customers.csv', index_col = False)

# assigning main project path to variable 'path'

path = r'C:\Users\Documents\Career Foundry\Data Immersion\Instacart Basket Analysis'

# 3. Data Wrangling

## Customer data characteristics

# displays the first 5 rows in data

customers.head()

# displays last 5 rows of dataframe

customers.tail()

# returns dimension of dataframe

customers.shape

# returns dataframe basic info

customers.info()

# returns data types in dataframe

customers.dtypes

# returns titles of columns

customers.columns

# 4. Renaming customer columns

# create a dictionary to rename multiple columns at once

dict = {'User ID' : 'user_id',
        'Surnam' : 'Surname',
       'Gender' : 'Sex',
       'STATE' : 'State',
       'date_joined' : 'Date Joined',
       'n_dependants' : 'Number Dependents',
       'fam_status' : 'Family Status',
       'income' : 'Income'}

# rename columns and save them

customers.rename(columns=dict,
                inplace=True)

# verify that changes were made

display(customers)

# 5. Checking and changing customers data types

# assigning new data types by creating a dictionary

convert_dict = {'user_id': 'str',
                'Age': np.int32,
                'Number Dependents': np.int8
               }

# overwriting old dtypes with new ones from dictionary

customers = customers.astype(convert_dict)

customers.dtypes

# returns descriptive statistics to verify that they make sense

customers.describe()

## Checking for mixed data types in customers df

# First Name column contains mixed data types

for col in customers.columns.tolist():
  weird = (customers[[col]].applymap(type) != customers[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (customers[weird]) > 0:
    print (col)

# Changing First Name column data type to string

customers['First Name'] = customers['First Name'].astype('str')

# verifying that change was made

for col in customers.columns.tolist():
  weird = (customers[[col]].applymap(type) != customers[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (customers[weird]) > 0:
    print (col)

# 6. Checking for nulls and duplicate values

# no null values in df
customers.isnull().sum()

# no duplicates found
df_dups = customers[customers.duplicated()]
df_dups

# checking shape of customers data frame after consistency checks
customers.shape

# 7. Exporting customers df

customers.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'customers.pkl'))

# 8. Import ords_prods_merge dataframe

# importing PKL from prepared data

ords_prods_merge = pd.read_pickle(r'C:\Users\kevan\Documents\Career Foundry\Data Immersion\Achievement 4\Instacart Basket Analysis\02 Data\Prepared Data\ords_prods_merge.pkl')

# 9. Addressing outliers in prices column

# Prices $14900 and 99999 are skewing results and are likely not correct
# Prices greater than $100 are labeled as NaN 

ords_prods_merge.loc[ords_prods_merge['prices'] >100, 'prices'] = np.nan

# 10. Change ords_prods_merge data types

# checking data types

ords_prods_merge.dtypes

# assigning new data types to ords_prods_merge by creating a dictionary

convert_dict2 = {'order_id': 'str',
                'user_id': 'str',
                'order_number': np.int32,
                'orders_day_of_week': np.int8,
                'order_hour_of_day': np.int8,
                'days_since_prior_order': np.float16,
                'product_id': 'str',
                'reordered': 'str',
                'aisle_id': 'str',
                'department_id': 'str',
                'prices': np.float32,
                'max_order': np.int32,
                'average_price': np.int32,
                'median_days_prior_order': np.float32
               }

# overwriting old dtypes with new ones from dictionary

ords_prods_merge = ords_prods_merge.astype(convert_dict2)
ords_prods_merge.dtypes

# exporting ords_prods_merge before merging with customers

ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_merge.pkl'))

# 11. Merging ords_prods_merge to customers

# display headers to find key column

ords_prods_merge.head(1)

# user_id is key column and is same data type as ords_prods_merge

customers.head(1)

# inner join ords_prods_merge onto customers with an indicator flag
# merged data frame called full_merge

full_merge = ords_prods_merge.merge(customers, on = ['user_id'], indicator = True)

# checking that merge was successful

full_merge.head()

# verify that all rows were successfully merged

full_merge['_merge'].value_counts()

#removing _merge flag column from final df

full_merge = ords_prods_merge.merge(customers, on = ['user_id'], indicator = False)

# verify that values are accurate. AKA max price is $25

full_merge.describe()

# 12. Exporting full_merge to pkl

full_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'full_merge.pkl'))
